Fix for regression - do not push upstream
This commit makes LNST always test both ipv6 and ipv4, regardless of what
specified in aliases. This way, the internal combined_queue branch can be
tested with ipv6 while the stable kernel does not.

This patch is for internal use and is not intended to be pushed to upstream
lnst versions

Signed-off-by: Yotam Gigi <yotamg@mellanox.com>