recipes: switchdev: ipip: Query XML network addresses
The IPIP tests need to set up a part of the overall network. The
numbering scheme has to be compatible with what's specified in the XML
network specification. So far, the dependency was implicit. Elevate the
aliases in each recipe's XML file to an API and look them up from Python
to make the dependency explicit.

Signed-off-by: Petr Machata <petrm@mellanox.com>