recipes: switchdev: ipip: Test conflict after VRF migration
Local conflict between tunnels arises when two tunnels use the same
local address in the same underlay VRF. That situation can come up in
several ways: on tunnel creation, as a result of a change, and when a
tunnel is moved to a VRF where another tunnel already has an underlay of
the same address.

The latter test was missing from the suite so far. Add it.

Structurally, this test is similar to the one that tests local conflict
due to a change in tunnel configuration. Link the two tests together in
comments, because any changes to one are likely to be interesting for
the other as well.

Also change the test for local conflict due to a change to require slow
path packets. The connection will work anyway: the offloaded tunnel will
just trap non-matching packets to slow path, where they are handled. But
to keep things consistent, it is desirable to have either both tunnels
in slow path, or none of them.

Signed-off-by: Petr Machata <petrm@mellanox.com>