recipes: switchdev: ipip-008: Move to default VRF
This recipe tests a situation that might lead to demotion of involved
tunnels to slow path. However, in slow path, the flat configuration
without bound device actually only works in the default VRF. So change
the recipe to execute the test there instead of another VRF--this ends
up testing the same code in the driver.

Signed-off-by: Petr Machata <petrm@mellanox.com>