recipes/switchdev: Correct ACL vlan modification
tc-001-action-vlan-modify.py is switching between different vlans using TC
on the switch, but the learning mechanism receives MACs that don't match
any of its known {port, vlan} as a result.

Notice this would not work on the SW model, and it manages to work
[albeit while logging errors] purely by chance due to the HW pipeline.

The solution is to add both vlans on both bridged ports.

Signed-off-by: Yuval Mintz <yuvalm@mellanox.com>