switchdev: recipes: Don't ping when both links are down
The LAG failover tests toggle the two LAG member links and make sure
ping passes even when only one link is active.

However, after we link up the first link we should wait for it to come
up before putting down the second link and starting the ping test.
Otherwise, we risk starting the test when both links are down.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Signed-off-by: Yuval Mintz <yuvalm@mellanox.com>