recipes: switchdev: ipip-006: Disable rp_filter
The decap-only test uses a highly-asymmetric configuration, where on the
switch, packets are forwarded without a corresponding encap route, and
on both hosts, responses (ICMP and ARP) are forwarded through a
different device than rx'd the request. For these reasons, rp_filter
needs to be disabled on all of those.

Signed-off-by: Petr Machata <petrm@mellanox.com>