recipes: switchdev: Basic RED test
Create a bottleneck for traffic by setting the ingress port to be in higher
rate the then egress port.
Set RED on the egress port as a step function (min = max - 1)

Check the behaviour in different traffic rates. Since it is not possible
to have complete control over the traffic (since it is dependent on the
sending host abilities to work without bursts), the rate is being modified
according to the backlog behavior.

The checks are:
* As long that the backlog is significantly lower then the step, there
  should be no drops.
* If the backlog is approximately the step value, there should be early
  drops.
* The RED stats should be coordinated with the rest of the stats.

Finally, RED is removed and there is a check to see that the traffic is
acting accordingly.

Signed-off-by: Nogah Frankel <nogahf@mellanox.com>
Reviewed-by: Petr Machata <petrm@mellanox.com>