recipes: switchdev: Basic RED-ECN test
Add a basic test for RED-ECN. It is very similar to the basic RED test,
but with some significant changes.
Packets are not dropped but marked. So the backlog can and should go over
the step limit in high rate. Pdrops are allowed in high rate and are
required in very high rate. For low rate we can have the limit on the
backlog size to be much closer to the step function.
Another difference is that checking that ECN was turned off must be checked
on the receiving host.

Signed-off-by: Nogah Frankel <nogahf@mellanox.com>
Reviewed-by: Petr Machata <petrm@mellanox.com>