Machine: clear netns from Slave's InterfaceManager
When deconfigure() is called for an Interface, the controller is sending
an RPC call to the slave on the netns where that interface is currently
at. This would lead to the slave changing the netns of the interface,
but the default InterfaceManager of the slave [the one of the parent
pid] would still have the interface marked as belonging to that netns.

Instead, the RPC should be sent without the netns indication, leading to
the parent processing it and clearing its own InterfaceManager DB.

Signed-off-by: Yuval Mintz <yuvalm@mellanox.com>