InterfaceManager: Be lenient when deleting routes
Upon cleanup it's possible that some of the configured routes are
no longer configured due to the teardown order of the controller.
Now that route cleanup is fixed, this may lead to the recipe failing as
some 'ip route del' command would fail on the slave.

To prevent the controller from having to try and account the effect each
such teardown may have on configured routes, simply be lenient on the
slave-side and accept failures to delete routes.

Signed-off-by: Yuval Mintz <yuvalm@mellanox.com>