switchdev: recipes: Increase number of generated flows in multipath recipe
The multipath recipe generates 10M flows and checks the Tx counters of
the nexthop devices to make sure the traffic was distributed according
to the provided weights.

Apparently, 10M is not enough and the test sometimes fails due to a
large discrepancy between the requested and measured ratios.

Increase the number of generated flows to 20M.

Fixes: 1ecdc72af238 ("switchdev: recipes: Add non-equal-cost multi-path recipe")
Signed-off-by: Ido Schimmel <idosch@mellanox.com>