XXX: Log exceptions
The fail that we are seeing comes from handling of MachineException.
There are only handful of places where that is raised, and fewer still
that are called repeatedly so that one could see them after some 500 or
so successful runs of the VLAN test: basically Interface.get_hwaddr,
Interface.get_devname and of course Machine.run_command when there's a
timeout. However simply causing a timeout does not reproduce the
original behavior, where we see _kill_command(None) failing to find the
command registered under the ID of None (yes, that appears to be valid).

So as a first step towards debugging this, let's dump the exception
thrown on the console, so that we have more information the next time
this triggers.

Obviously this is not intended for upstream.