recipes: switchdev: Test RED on two children
Create a bottleneck for traffic by setting the ingress port to be in higher
rate the then egress port.
Set PRIO with 2 bands (3 are configured but only 2 are used).
Set on each band RED with different configuration.
Test each band in the same way that the basic RED test does.
Send the traffic to the wanted band by changing its TOS field.

Signed-off-by: Nogah Frankel <nogahf@mellanox.com>
Reviewed-by: Petr Machata <petrm@mellanox.com>