recipes: switchdev: sharedbuffer: Support pool 8
Pool #8 was recently introduced to mlxsw to support MC-aware regime of
the switch. Pool 8 is configured with infinite size, but unfortunately
devlink does not support reporting infinites. So instead, the pool size
is reported as the full switch memory, which however cannot be
configured explicitly.

Tweak the sharedbuffer test to skip the MC pool 8, so that it does not
fail when attempting to reconfigure the reported size.

Signed-off-by: Petr Machata <petrm@mellanox.com>