recipes: switchdev: ipip-007: Work around driver changes
In kernel commit fd74feeb2452 ("Add layer 3 devlink-trap exceptions
support"), packets trapped with IPV4_LPM_UNICAST_MISS are configured to be
trapped, instead of dropped. As a result, some of the ICMP packets that
this test uses are trapped. Linux then forwards them, because there is a
behavior mismatch between the offloaded datapath and the Linux one.
So the ping passes, causing the test to fail.

But the lack of conflict can be checked by the ICMP6 test above, which is
required to go in offloaded data path. Requiring the ICMP test to go in
Linux path checks the assumption that the offloaded data path punted the
packet to the slow path. The previous check was that the offloaded data
path dropped the packet.

Signed-off-by: Petr Machata <petrm@mellanox.com>