recipes: switchdev: Update IPIP tests for the kernel fix
The mlxsw driver used to assume that a GRE netdevice without a bind device
has as underlay the same VRF that the GRE netdevice is in. In Linux
however, the underlay in such case is the main VRF. That discrepancy was
fixed recently. Update some test cases to work well with the new
configuration.

In ipip-003-gre-flat, move the underlay forwarding route, the dummy tunnel
endpoint, and the underlay netdevice from VRF1 to the main VRF.

ipip-007-gre-migration tests a conflict in local address caused by moving a
GRE device to another VRF. However since the fix, it is not possible to
cause a conflict by moving an unbound GRE tunnel. So just drop the test.

Signed-off-by: Petr Machata <petrm@mellanox.com>